+++ src/components/footer/footer.tsx
@@ -1,7 +1,10 @@
+import {Link} from 'react-router-dom';
+import {RoutePath} from '../../routes';
+
 function Footer(): JSX.Element {
   return (
     <footer className="footer container">
-      <a className="footer__logo-link" href="main.html">
+      <Link className="footer__logo-link" to={RoutePath.INDEX}>
         <img
           className="footer__logo"
           src="/img/logo.svg"
@@ -9,7 +12,7 @@ function Footer(): JSX.Element {
           width={64}
           height={33}
         />
-      </a>
+      </Link>
     </footer>
   );
 }
