+++ src/store/action.ts
@@ -2,4 +2,4 @@ import {createAction} from '@reduxjs/toolkit';
 import {CityName, SortType} from '../types';
 
 export const setActiveCityName = createAction<CityName>('main/setActiveCityName');
-export const changeOfferSortOption = createAction<SortType>('main/changeOfferSortOption');
+export const setOfferSortOption = createAction<SortType>('main/changeOfferSortOption');
